Create README.TXT
Quick description without details. Would be good to expand on this in a good way to make sure that it can be used to explain all the functions and sub-sites available from the main site. Maybe even divide up into several different README files to make it easier to understan what everything is for